CXProgram.go: Start on making the code more object oriented.
This commit extracts the definition of and all methods on the struct CXProgram
into a new file called CXProgram.go.

This is a first commit in a series that aims at making CX implemented in a
amore object oriented fashion.

The commit does not touch serialization, execution or affordances.  Those may
be later, after we have had some discussion about it.