Alternate approach for sending the request
Since there's a issue with propagating req struct from cx
there was a change in the approach on how to make the request

http.NewRequest is now used to prepare the request object and to invoke
client to send it over.

Existing issue is documented in the op_http.go on line with following:
TODO issue with returning response
Issue is documented and happens when proper response is serialized.